// app/admin/(routes)/payments/types.ts
export interface CreateWalletResponseData {
  address_in: string
  polygon_address_in: string
  callback_url: string
  ipn_token: string
}

export type FiatCurrency =
  | 'USD'
  | 'EUR'
  | 'CAD'
  | 'GBP'
  | 'AUD'
  | 'JPY'
  | 'CHF'
  | 'CNY'
  | 'INR'
  | 'BRL'
  | 'MXN'
  | 'SGD'
  | 'HKD'
  | 'NZD'
  | 'ZAR'
  | 'SEK'
  | 'NOK'
  | 'DKK'
  | 'PLN'
  | 'TRY'
  | 'PHP'

export type Fiat =
  | 'USD'
  | 'EUR'
  | 'GBP'
  | 'CAD'
  | 'JPY'
  | 'AED'
  | 'MYR'
  | 'IDR'
  | 'THB'
  | 'CHF'
  | 'SGD'
  | 'RUB'
  | 'ZAR'
  | 'TRY'
  | 'LKR'
  | 'RON'
  | 'BGN'
  | 'HUF'
  | 'CZK'
  | 'PHP'
  | 'PLN'
  | 'UGX'
  | 'MXN'
  | 'INR'
  | 'HKD'
  | 'CNY'
  | 'BRL'
  | 'DKK'
  | 'TWD'
  | 'AUD'
  | 'NGN'
  | 'SEK'
  | 'NOK'
  | 'UAH'

export type CryptoTicker =
  | 'btc'
  | 'bch'
  | 'ltc'
  | 'doge'
  | 'eth'
  | 'trx'
  | 'bep20/1inch'
  | 'bep20/ada'
  | 'bep20/bnb'
  | 'bep20/btcb'
  | 'bep20/cake'
  | 'bep20/dai'
  | 'bep20/doge'
  | 'bep20/eth'
  | 'bep20/ltc'
  | 'bep20/phpt'
  | 'bep20/shib'
  | 'bep20/usd1'
  | 'bep20/usdc'
  | 'bep20/usdt'
  | 'bep20/xrp'
  | 'erc20/1inch'
  | 'erc20/arb'
  | 'erc20/bnb'
  | 'erc20/cbbtc'
  | 'erc20/dai'
  | 'erc20/link'
  | 'erc20/ondo'
  | 'erc20/pepe'
  | 'erc20/pol'
  | 'erc20/pyusd'
  | 'erc20/shib'
  | 'erc20/usd1'
  | 'erc20/usdc'
  | 'erc20/usdp'
  | 'erc20/usdt'
  | 'erc20/wxrp'
  | 'arbitrum/arb'
  | 'arbitrum/dai'
  | 'arbitrum/eth'
  | 'arbitrum/link'
  | 'arbitrum/pepe'
  | 'arbitrum/usdc'
  | 'arbitrum/usdc.e'
  | 'arbitrum/usdt'
  | 'avax-c/avax'
  | 'avax-c/usdc'
  | 'avax-c/usdc.e'
  | 'avax-c/usdt'
  | 'avax-c/wavax'
  | 'avax-c/weth.e'
  | 'polygon/avax'
  | 'polygon/pol'
  | 'polygon/usdc'
  | 'polygon/usdc.e'
  | 'polygon/usdt'
  | 'polygon/weth'
  | 'base/cbbtc'
  | 'base/dai'
  | 'base/eth'
  | 'base/usdc'
  | 'base/usdt'
  | 'optimism/dai'
  | 'optimism/eth'
  | 'optimism/link'
  | 'optimism/op'
  | 'optimism/usdc'
  | 'optimism/usdc.e'
  | 'optimism/usdt'
  | 'sol/cbbtc'
  | 'sol/eurc'
  | 'sol/sol'
  | 'sol/trump'
  | 'sol/usdc'
  | 'sol/usdt'
  | 'sol/wbtc'
  | 'sol/weth'
  | 'trc20/btc'
  | 'trc20/inrt'
  | 'trc20/tusd'
  | 'trc20/usdd'
  | 'trc20/usdt'